Replace float by decimal

+++ expert_dollup/app/dtos/report_definition_dto.py
@@ -1,8 +1,9 @@
 from uuid import UUID
 from typing import List, Dict, Union, Optional
+from decimal import Decimal
 from expert_dollup.shared.starlette_injection import CamelModel
 
-ReportColumnDictDto = Dict[str, Union[str, float, bool, int, UUID, List[UUID], None]]
+ReportColumnDictDto = Dict[str, Union[str, Decimal, bool, int, UUID, List[UUID], None]]
 ReportRowDictDto = Dict[str, ReportColumnDictDto]
